requirements.txt
requests==2.25.1
geopy==2.2.0
pyfiglet==0.8.post1
pandas==1.3.2
folium==0.12.1
openpyxl==3.0.7
beautifulsoup4==4.10.0
Pillow==8.4.0
PyPDF2==1.26.0
